Add files via upload
This is a login page connected with a signup page made with the help of tkinter purely in python. A user can fill his/her username and password in it if he/she is already registered , if not there is an option to sign up as being a new user by clicking upon the signup text. After clicking the user will be redirected to another page of registration where he/she have to enter his name password and confirm it, and all of the filled details will be saved into notepad datasheet.